+++ src/main/java/kitchenpos/domain/menu/MenuProduct.java
@@ -8,6 +8,8 @@ import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.ManyToOne;
 
+import kitchenpos.domain.product.Product;
+
 @Entity
 public class MenuProduct {
