+++ src/main/java/kitchenpos/domain/menugroup/MenuGroupRepository.java
@@ -1,8 +1,8 @@
-package kitchenpos.domain.menu;
+package kitchenpos.domain.menugroup;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import kitchenpos.domain.menu.MenuGroup;
+import kitchenpos.domain.menugroup.MenuGroup;
 
 public interface MenuGroupRepository extends JpaRepository<MenuGroup, Long> {
 }
