+++ src/main/java/kitchenpos/domain/order/OrderRepository.java
@@ -1,10 +1,12 @@
-package kitchenpos.domain;
+package kitchenpos.domain.order;
 
 import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
+
+import kitchenpos.domain.order.Order;
 
 public interface OrderRepository extends JpaRepository<Order, Long> {
