+++ src/main/java/kitchenpos/domain/product/ProductRepository.java
@@ -1,6 +1,8 @@
-package kitchenpos.domain.menu;
+package kitchenpos.domain.product;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+
+import kitchenpos.domain.product.Product;
 
 public interface ProductRepository extends JpaRepository<Product, Long> {
