+++ src/main/java/kitchenpos/domain/table/OrderTableRepository.java
@@ -1,10 +1,12 @@
-package kitchenpos.domain;
+package kitchenpos.domain.table;
 
 import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
+
+import kitchenpos.domain.table.OrderTable;
 
 public interface OrderTableRepository extends JpaRepository<OrderTable, Long> {
