+++ src/main/java/kitchenpos/domain/tablegroup/TableGroupRepository.java
@@ -1,6 +1,8 @@
-package kitchenpos.domain;
+package kitchenpos.domain.tablegroup;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+
+import kitchenpos.domain.tablegroup.TableGroup;
 
 public interface TableGroupRepository extends JpaRepository<TableGroup, Long> {
 }
