+++ src/main/java/kitchenpos/dto/OrderRequest.java
@@ -3,10 +3,10 @@ package kitchenpos.dto;
 import java.util.List;
 import java.util.stream.Collectors;
 
-import kitchenpos.domain.Order;
-import kitchenpos.domain.OrderLineItem;
-import kitchenpos.domain.OrderLineItems;
-import kitchenpos.domain.OrderTable;
+import kitchenpos.domain.order.Order;
+import kitchenpos.domain.order.OrderLineItem;
+import kitchenpos.domain.order.OrderLineItems;
+import kitchenpos.domain.table.OrderTable;
 
 public class OrderRequest {
     private Long orderTableId;
