+++ src/main/java/kitchenpos/dto/OrderTableRequest.java
@@ -1,6 +1,6 @@
 package kitchenpos.dto;
 
-import kitchenpos.domain.OrderTable;
+import kitchenpos.domain.table.OrderTable;
 
 public class OrderTableRequest {
     private int numberOfGuests;
