+++ src/main/java/kitchenpos/ui/ProductRestController.java
@@ -1,7 +1,7 @@
 package kitchenpos.ui;
 
 import kitchenpos.application.ProductService;
-import kitchenpos.domain.menu.Product;
+import kitchenpos.domain.product.Product;
 import kitchenpos.dto.ProductRequest;
 
 import org.springframework.http.ResponseEntity;
