+++ src/main/java/kitchenpos/ui/TableGroupRestController.java
@@ -1,7 +1,7 @@
 package kitchenpos.ui;
 
 import kitchenpos.application.TableGroupService;
-import kitchenpos.domain.TableGroup;
+import kitchenpos.domain.tablegroup.TableGroup;
 import kitchenpos.dto.TableGroupRequest;
 
 import org.springframework.http.ResponseEntity;
